requirements.txt
agate==1.9.1
alembic==1.17.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.14.2
bleach==6.2.0
boto3==1.40.47
botocore==1.40.47
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==14.0
comm==0.2.3
contourpy==1.3.3
copulas==0.12.3
ctgan==0.11.0
cycler==0.12.1
daff==1.4.2
dagster==1.11.16
dagster-airbyte==0.27.16
dagster-dbt==0.27.16
dagster-graphql==1.11.16
dagster-pipes==1.11.16
dagster_shared==1.11.16
dagster-webserver==1.11.16
dbt-adapters==1.17.3
dbt-common==1.33.0
dbt-core==1.9.10
dbt-extractor==0.6.0
dbt-postgres==1.9.1
dbt-protos==1.0.382
dbt-semantic-interfaces==0.7.5
debugpy==1.8.17
decorator==5.2.1
deepdiff==8.6.1
deepecho==0.7.0
defusedxml==0.7.1
diff_cover==9.7.1
docstring_parser==0.17.0
executing==2.2.1
Faker==37.11.0
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.76.0
grpcio-health-checking==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
kiwisolver==1.4.9
lark==1.3.0
leather==0.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
narwhals==2.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
ordered-set==4.1.0
orderly-set==5.5.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.5
pathlib_abc==0.5.2
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
psycopg2-binary==2.9.11
psycopg==3.2.12
psycopg-binary==3.2.12
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rdt==1.18.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
sdmetrics==0.23.0
sdv==1.20.0
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlfluff==2.3.5
sqlglot==27.28.1
sqlglotrs==0.7.3
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.49.0
structlog==25.5.0
sympy==1.14.0
tabulate==0.9.0
tblib==3.2.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.3.0
tomlkit==0.13.3
toposort==1.10
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typer==0.20.0
types-python-dateutil==2.9.0.20251008
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
universal_pathlib==0.3.4
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.14
yarl==1.22.0
zipp==3.23.0
